dailyfresh/requirements.txt
amqp==1.4.9
anyjson==0.3.3
asgiref==3.2.10
billiard==3.3.0.23
celery==3.1.26.post2
cffi==1.14.3
cryptography==3.2.1
Django==1.11.16
django-celery==3.3.1
django-haystack==2.8.0
django-redis==4.12.1
djangorestframework==3.11.1
drf-haystack==1.8.9
importlib-metadata==1.7.0
itsdangerous==1.1.0
kombu==3.0.37
Pillow==8.0.0
pycparser==2.20
pycryptodomex==3.9.4
PyMySQL==0.10.0
pyOpenSSL==19.1.0
python-alipay-sdk==2.1.1
python-dateutil==2.8.1
pytz==2020.1
redis==3.5.3
six==1.15.0
sqlparse==0.4.1
vine==1.3.0
Whoosh==2.7.4
zipp==3.1.0
